Remove onNavForward for newCondition last page; move clear utils to index
